fixes for travis, clang++

+++ lib-http/src/uri.cpp
@@ -109,7 +109,7 @@ std::filesystem::path uri::get_path() const
 	std::filesystem::path result;
 
 	for (auto p = m_impl->m_uri.pathHead; p != nullptr /*m_impl->m_uri.pathTail*/; p = p->next)
-		result /= decode_url({ p->text.first, p->text.afterLast - p->text.first });
+		result /= decode_url({ p->text.first, static_cast<size_t>(p->text.afterLast - p->text.first) });
 	
 	return result;
 }
